+++ lib/money2_fixer.dart
@@ -1,3 +1,1 @@
-library money2_fixer;
-
 export 'src/money2_fixer_ext.dart';