+++ test/fixed_test.dart
@@ -1,5 +1,5 @@
-import 'package:test/test.dart';
 import 'package:money2/money2.dart';
+import 'package:test/test.dart';
 
 void main() {
   const maxScale = 100;
@@ -9,17 +9,23 @@ void main() {
     for (var scale = 0; scale <= maxScale; scale++) {
       final str = scale == 0 ? '1' : '1.${'0' * (scale - 1)}1';
       final fmt = scale == 0 ? 'S#' : '#.${'#' * scale}';
-      expect(Fixed.parse(str, scale: scale).format(fmt), str,
-          reason: 'Failed with $scale scale');
+      expect(
+        Fixed.parse(str, scale: scale).format(fmt),
+        str,
+        reason: 'Failed with $scale scale',
+      );
     }
   });
 
   test('integers 0-$maxInts test', () {
     for (var ints = 0; ints <= maxInts; ints++) {
       final str = ints == 0 ? '0' : '9' * ints;
-      final fmt = '#';
-      expect(Fixed.parse(str, scale: 0).format(fmt), str,
-          reason: 'Failed with $ints ints');
+      const fmt = '#';
+      expect(
+        Fixed.parse(str, scale: 0).format(fmt),
+        str,
+        reason: 'Failed with $ints ints',
+      );
     }
   });
 
@@ -36,8 +42,11 @@ void main() {
 
         final fmt = scale == 0 ? '0' : '0.${'0' * scale}';
 
-        expect(Fixed.parse(str, scale: scale).format(fmt), expectStr,
-            reason: 'Failed with $scale scale, $ints ints');
+        expect(
+          Fixed.parse(str, scale: scale).format(fmt),
+          expectStr,
+          reason: 'Failed with $scale scale, $ints ints',
+        );
       }
     }
   });