+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 3.0.0
+
+* Fix broken by money2 toJson and fromJson.
+* Breaking: add toJsonImproved and fromJsonImproved instead of toJson and fromJson.
+* Add very good analysis lints.
+
 ## 2.0.0
 
 * Bump money2 version to ^5.2.1.